requirements.txt
langchain
langgraph
streamlit
pandas
python-dotenv
langchain-google-genai==3.0.0
requests
